refactor(driver-api): modularize sql_dump module
Split monolithic sql_dump.rs into dump/restore/parser submodules with
mod.rs facade re-exports for zero breaking changes to callers.
